requirements.txt
pysam==0.19.1
pyvcf==0.6.8
